refactor(cockpit): extract project settings app
