+++ src/pollypm/cockpit_project_settings.py
@@ -1,0 +1,195 @@
+"""Cockpit project settings panel.
+
+Contract:
+- Inputs: a cockpit config path and project key.
+- Outputs: ``PollyProjectSettingsApp`` for project session/account control.
+- Side effects: loads config, inspects project sessions, and issues
+  service calls to stop or retarget the project's worker session.
+- Invariants: rendering and behavior stay local to this screen; callers
+  can continue importing the app from ``pollypm.cockpit_ui`` for
+  compatibility while the implementation lives here.
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+
+from textual import on
+from textual.app import App, ComposeResult
+from textual.containers import Horizontal, Vertical
+from textual.widgets import Button, Static
+
+from pollypm.config import load_config
+from pollypm.models import ProviderKind
+from pollypm.service_api import PollyPMService
+
+
+class PollyProjectSettingsApp(App[None]):
+    TITLE = "PollyPM"
+    SUB_TITLE = "Project Settings"
+    CSS = """
+    Screen {
+        background: #0c0f12;
+        color: #eef2f4;
+        padding: 1;
+    }
+    #title-bar {
+        height: 1;
+        color: #5b8aff;
+        text-style: bold;
+        padding-bottom: 1;
+    }
+    #message {
+        height: 1;
+        color: #7ee8a4;
+        padding-bottom: 1;
+    }
+    .settings-section {
+        padding: 1;
+        border: round #253140;
+        background: #111820;
+        margin-bottom: 1;
+    }
+    .section-label {
+        color: #5b8aff;
+        text-style: bold;
+        padding-bottom: 1;
+    }
+    .field-row {
+        height: auto;
+        padding-bottom: 1;
+    }
+    .field-label {
+        color: #6b7a88;
+        width: 12;
+    }
+    .field-value {
+        color: #e0e8ef;
+    }
+    #actions {
+        height: auto;
+        padding-top: 1;
+    }
+    #actions Button {
+        margin-right: 1;
+    }
+    """
+
+    def __init__(self, config_path: Path, project_key: str) -> None:
+        super().__init__()
+        self.config_path = config_path
+        self.project_key = project_key
+        self.title_bar = Static("", id="title-bar")
+        self.message_bar = Static("", id="message")
+
+    def compose(self) -> ComposeResult:
+        yield self.title_bar
+        yield self.message_bar
+        with Vertical(classes="settings-section"):
+            yield Static("Worker Session", classes="section-label")
+            yield Static("", id="worker-info")
+        with Vertical(classes="settings-section"):
+            yield Static("Model & Account", classes="section-label")
+            yield Static("", id="model-info")
+        with Horizontal(id="actions"):
+            yield Button("Reset Session", id="reset-session", variant="warning")
+            yield Button("Switch to Claude", id="switch-claude", variant="primary")
+            yield Button("Switch to Codex", id="switch-codex", variant="primary")
+
+    def on_mount(self) -> None:
+        self._refresh()
+
+    def _refresh(self) -> None:
+        config = load_config(self.config_path)
+        project = config.projects.get(self.project_key)
+        if project is None:
+            self.title_bar.update(f"Project not found: {self.project_key}")
+            return
+        self.title_bar.update(f"{project.name or project.key} • Settings")
+
+        worker = None
+        for session in config.sessions.values():
+            if session.role == "worker" and session.project == self.project_key and session.enabled:
+                worker = session
+                break
+
+        worker_info = self.query_one("#worker-info", Static)
+        model_info = self.query_one("#model-info", Static)
+
+        if worker is None:
+            worker_info.update("No worker session configured.\nPress N in the sidebar to create one.")
+            model_info.update("")
+            return
+
+        account = config.accounts.get(worker.account)
+        account_label = f"{account.email} [{account.provider.value}]" if account else worker.account
+        worker_info.update(
+            f"[dim]Session:[/] [bold]{worker.name}[/]\n"
+            f"[dim]Window:[/]  {worker.window_name}\n"
+            f"[dim]CWD:[/]     {worker.cwd}"
+        )
+        model_info.update(
+            f"[dim]Provider:[/] [bold]{worker.provider.value}[/]\n"
+            f"[dim]Account:[/]  {account_label}\n"
+            f"[dim]Args:[/]     {' '.join(worker.args) if worker.args else 'none'}"
+        )
+
+    def _notify(self, msg: str) -> None:
+        self.message_bar.update(msg)
+
+    @on(Button.Pressed, "#reset-session")
+    def on_reset(self, event: Button.Pressed) -> None:
+        del event
+        config = load_config(self.config_path)
+        worker = None
+        for session in config.sessions.values():
+            if session.role == "worker" and session.project == self.project_key and session.enabled:
+                worker = session
+                break
+        if worker is None:
+            self._notify("No worker session to reset.")
+            return
+        try:
+            PollyPMService(self.config_path).stop_session(worker.name)
+            self._notify(f"Session {worker.name} stopped. Press N to relaunch.")
+        except Exception as exc:  # noqa: BLE001
+            self._notify(f"Reset failed: {exc}")
+        self._refresh()
+
+    @on(Button.Pressed, "#switch-claude")
+    def on_switch_claude(self, event: Button.Pressed) -> None:
+        del event
+        self._switch_provider(ProviderKind.CLAUDE)
+
+    @on(Button.Pressed, "#switch-codex")
+    def on_switch_codex(self, event: Button.Pressed) -> None:
+        del event
+        self._switch_provider(ProviderKind.CODEX)
+
+    def _switch_provider(self, target_provider: ProviderKind) -> None:
+        config = load_config(self.config_path)
+        worker = None
+        for session in config.sessions.values():
+            if session.role == "worker" and session.project == self.project_key and session.enabled:
+                worker = session
+                break
+        if worker is None:
+            self._notify("No worker session to switch.")
+            return
+        target_account = None
+        for name, account in config.accounts.items():
+            if account.provider is target_provider:
+                target_account = name
+                break
+        if target_account is None:
+            self._notify(f"No {target_provider.value} account available.")
+            return
+        if worker.provider is target_provider:
+            self._notify(f"Already using {target_provider.value}.")
+            return
+        try:
+            PollyPMService(self.config_path).switch_session_account(worker.name, target_account)
+            self._notify(f"Switched to {target_provider.value} ({target_account}). Session restarted.")
+        except Exception as exc:  # noqa: BLE001
+            self._notify(f"Switch failed: {exc}")
+        self._refresh()
